# docs/openapi.yaml
openapi: 3.0.3
info:
  title: API de Órdenes de Compra a Proveedores
  description: >
    Contrato de referencia para la implementación del backend. Este archivo es
    el contrato preciso contra el cual se implementa y se mantiene a mano.
  version: "2026-09-24"
  contact:
    name: Henry Molina
    url: https://henry-molina.dev/
servers:
  - url: http://localhost:8080/api
    description: Entorno local (Docker Compose)
security:
  - bearerAuth: []
...
paths:
...
  /ordenes:
    get:
      tags: [Ordenes]
      summary: Listar órdenes de compra
      operationId: listarOrdenes
      description: >
        Vista central de trabajo de Compras: todas las órdenes visibles para el
        usuario autenticado, con filtros y paginación por cursor. Un
        GERENTE_SUCURSAL ve únicamente las órdenes de su propia sucursal
        (filtrado automático por el backend según su usuario, no por parámetro
        del cliente); ADMIN y COMPRADOR ven todas.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: estado
          in: query
          schema: { $ref: "#/components/schemas/EstadoOrden" }
        - name: proveedor_id
          in: query
          schema: { type: integer }
        - name: sucursal_id
          in: query
          schema: { type: integer }
        - name: cliente_id
          in: query
          schema: { type: integer }
        - name: numero_orden
          in: query
          schema: { type: string }
          description: >
            Busca coincidencia parcial, sin distinguir mayúsculas/minúsculas
            (ej. "123" encuentra "OC-2026-000123").
        - name: fecha_desde
          in: query
          schema: { type: string, format: date }
          description: Filtra por fecha_necesaria >= este valor. Se puede usar sola o junto a fecha_hasta.
        - name: fecha_hasta
          in: query
          schema: { type: string, format: date }
          description: Filtra por fecha_necesaria <= este valor. Se puede usar sola o junto a fecha_desde.
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/StartingAfter"
        - $ref: "#/components/parameters/Expand"
      responses:
        "200":
          description: Listado paginado de órdenes
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/OrdenCompra" }
        "400":
          description: fecha_desde posterior a fecha_hasta
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    post:
      tags: [Ordenes]
      summary: Crear orden de compra (estado inicial CREADA)
      operationId: crearOrden
      description: >
        Punto de partida del flujo de Compras: registra una nueva orden con su
        detalle de productos, valida las reglas de negocio (formato de sucursal
        vs. categoría permitida, cliente obligatorio en Venta Directa) y calcula
        el total. Disponible para COMPRADOR, que es quien arma órdenes en el
        día a día, y ADMIN. Un GERENTE_SUCURSAL no crea órdenes: solo las recibe
        y las cierra.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrdenCompraRequest" }
      responses:
        "201":
          description: Orden creada
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrdenCompra" }
        "400":
          $ref: "#/components/responses/InvalidRequestError"
  /ordenes/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPathParam"
    get:
      tags: [Ordenes]
      summary: Detalle de una orden de compra
      operationId: obtenerOrden
      description: >
        Cabecera y detalle completos de una orden puntual. Mismo criterio de
        visibilidad que el listado: un GERENTE_SUCURSAL solo puede consultar
        órdenes de su propia sucursal.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Expand"
      responses:
        "200":
          description: Detalle de la orden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrdenCompra" }
        "403":
          description: GERENTE_SUCURSAL intentando acceder a una orden de otra sucursal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          $ref: "#/components/responses/NotFoundError"
    put:
      tags: [Ordenes]
      summary: Modificar el detalle de una orden (solo si está en estado CREADA)
      operationId: actualizarOrden
      description: >
        Corrige las líneas de una orden antes de que avance en su ciclo de
        vida; una vez aprobada, anulada o cerrada ya no se puede editar (ver
        respuesta 409). Mismos roles que pueden crear una orden: COMPRADOR y
        ADMIN.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrdenCompraRequest" }
      responses:
        "200":
          description: Orden actualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrdenCompra" }
        "409":
          $ref: "#/components/responses/StateConflictError"
  /ordenes/{id}/aprobar:
    parameters:
      - $ref: "#/components/parameters/IdPathParam"
    patch:
      tags: [Ordenes]
      summary: Transición CREADA → APROBADA (dispara notificación al proveedor)
      operationId: aprobarOrden
      description: >
        Confirma la orden y dispara el correo de notificación al proveedor;
        solo es válida desde el estado CREADA (ver respuesta 409). La ejecuta
        quien la creó en primer lugar: COMPRADOR o ADMIN.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Orden aprobada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrdenCompra" }
        "409":
          $ref: "#/components/responses/StateConflictError"
  /ordenes/{id}/anular:
    parameters:
      - $ref: "#/components/parameters/IdPathParam"
    patch:
      tags: [Ordenes]
      summary: Transición a ANULADA (requiere motivo; no permitida desde CERRADA)
      operationId: anularOrden
      description: >
        Cancela una orden que ya no se necesita, dejando registrado el motivo
        para auditoría. No es una acción destructiva a nivel de datos (la orden
        queda anulada, no se borra) y no se puede aplicar sobre una orden ya
        CERRADA. Misma autorización que crearla y aprobarla: COMPRADOR o ADMIN.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [motivo]
              properties:
                motivo: { type: string, minLength: 1, maxLength: 250 }
      responses:
        "200":
          description: Orden anulada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrdenCompra" }
        "409":
          $ref: "#/components/responses/StateConflictError"
  /ordenes/{id}/cerrar:
    parameters:
      - $ref: "#/components/parameters/IdPathParam"
    patch:
      tags: [Ordenes]
      summary: Transición APROBADA → CERRADA (confirma recepción)
      operationId: cerrarOrden
      description: >
        Cierra el ciclo de vida de la orden confirmando si la mercancía llegó
        conforme o no. Solo válida desde APROBADA (ver respuesta 409). La
        ejecuta quien recibe físicamente el pedido: el GERENTE_SUCURSAL de la
        sucursal destino de la orden (nunca de otra), o ADMIN. Un COMPRADOR no
        puede cerrarla — no es quien recibe la mercancía.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [conforme]
              properties:
                conforme: { type: boolean }
                observacion: { type: string, maxLength: 250 }
      responses:
        "200":
          description: Orden cerrada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrdenCompra" }
        "403":
          description: Un COMPRADOR, o un GERENTE_SUCURSAL de otra sucursal, no puede ejecutar esta transición
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          $ref: "#/components/responses/StateConflictError"
  /ordenes/{id}/linea-tiempo:
    parameters:
      - $ref: "#/components/parameters/IdPathParam"
    get:
      tags: [Ordenes]
      summary: Historial combinado (auditoría interna + eventos de proveedor), ordenado por fecha
      operationId: obtenerLineaTiempoOrden
      description: >
        Vista de seguimiento pensada para que quien sigue una orden entienda su
        ciclo de vida completo de un vistazo, sin cruzar dos endpoints
        distintos: combina los cambios de estado internos con los eventos
        logísticos que reporta el proveedor. Mismos roles y misma restricción
        de sucursal para GERENTE_SUCURSAL que ver el detalle de la orden.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Línea de tiempo de la orden
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/LineaTiempoItem" }
...
  /ordenes/importaciones/plantilla:
    get:
      tags: [Ordenes]
      summary: Descargar plantilla Excel para carga masiva de órdenes
      operationId: descargarPlantillaImportacion
      description: >
        Punto de partida para una carga masiva: un .xlsx de ejemplo con la
        hoja "Ordenes" (columnas exactas que espera el endpoint de
        importación; la última, precio_unitario, es opcional: vacía usa el
        precio de catalogo, y con valor se valida contra los limites de
        negociacion del proveedor igual que en el flujo individual) más
        hojas de referencia (Proveedores/Sucursales/
        Clientes/Productos) con listas desplegables, para que el usuario
        elija un nombre válido en vez de tener que conocer un ID interno.
        Mismos roles que pueden crear órdenes de forma individual, ya que la
        carga masiva reutiliza exactamente el mismo servicio y las mismas
        reglas de negocio.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Archivo .xlsx de plantilla
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
  /ordenes/importaciones:
    post:
      tags: [Ordenes]
      summary: Carga masiva de órdenes de compra vía Excel
      operationId: importarOrdenes
      description: >
        Registra varias órdenes de una sola vez para casos de reposición
        periódica, pedidos recurrentes o migración de datos: agrupa las filas
        de la hoja "Ordenes" del archivo .xlsx por referencia_lote, resuelve
        proveedor/sucursal/cliente por nombre y producto por código, arma el
        mismo OrdenCompraRequest que el flujo individual y reutiliza el mismo
        servicio de creación de orden, así que ninguna regla de negocio se
        duplica ni se relaja. Atomicidad por orden (lote), no por archivo
        completo: un lote inválido se rechaza sin afectar a los demás. Mismos
        roles que crear una orden individual.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [archivo]
              properties:
                archivo:
                  type: string
                  format: binary
      responses:
        "200":
          description: Reporte de resultado de la importación
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportacionLoteResponse" }
  /ordenes/{id}/exportar:
    parameters:
      - $ref: "#/components/parameters/IdPathParam"
    get:
      tags: [Ordenes]
      summary: Exportar una orden de compra como PDF
      operationId: exportarOrdenPdf
      description: >
        Genera el documento formal de la orden (cabecera + detalle) para
        enviar al proveedor, imprimir o archivar. Mismo control de acceso que
        consultar la orden: GERENTE_SUCURSAL restringido a su propia sucursal.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: formato
          in: query
          required: true
          schema: { type: string, enum: [pdf] }
      responses:
        "200":
          description: Documento PDF de la orden (cabecera + detalle)
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "403":
          description: GERENTE_SUCURSAL intentando exportar una orden de otra sucursal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /ordenes/exportar:
    get:
      tags: [Ordenes]
      summary: Exportar un reporte Excel de órdenes de compra
      operationId: exportarOrdenesExcel
      description: >
        Reporte tabular para análisis o conciliación externa, con los mismos
        filtros que el listado. Una fila por línea de producto (no por
        orden), con los datos de cabecera repetidos en cada fila, para no
        perder el detalle de qué se compró. Mismos roles y misma restricción
        de sucursal para GERENTE_SUCURSAL que el listado de órdenes.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: formato
          in: query
          required: true
          schema: { type: string, enum: [xlsx] }
        - name: estado
          in: query
          schema: { $ref: "#/components/schemas/EstadoOrden" }
        - name: proveedor_id
          in: query
          schema: { type: integer }
        - name: sucursal_id
          in: query
          schema: { type: integer }
        - name: cliente_id
          in: query
          schema: { type: integer }
        - name: numero_orden
          in: query
          schema: { type: string }
        - name: fecha_desde
          in: query
          schema: { type: string, format: date }
        - name: fecha_hasta
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Reporte .xlsx
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
        "400":
          description: fecha_desde posterior a fecha_hasta
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT propio emitido por /auth/login (sin OAuth2)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Credencial por proveedor, usada exclusivamente por el endpoint de webhooks
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      schema: { type: integer }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: Clave generada por el cliente para evitar reprocesar la misma operación ante reintentos
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
    StartingAfter:
      name: starting_after
      in: query
      schema: { type: string }
      description: Cursor de paginación (número de orden del último elemento visto)
    Expand:
      name: expand
      in: query
      schema:
        type: array
        items: { type: string }
      style: form
      explode: true
      description: "Relaciones a embeber en la respuesta, ej. ?expand=proveedor&expand=linea_tiempo"
  responses:
    InvalidRequestError:
      description: Error de validación de entrada
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    NotFoundError:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    StateConflictError:
      description: Transición de estado inválida para el estado actual de la orden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    EstadoOrden:
      type: string
      enum: [CREADA, APROBADA, ANULADA, CERRADA]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [type, code, message]
          properties:
            type:
              type: string
              enum: [invalid_request_error, authentication_error, state_conflict_error, api_error]
            code: { type: string, example: categoria_no_permitida }
            message: { type: string }
            param: { type: string, nullable: true }
    ListEnvelope:
      type: object
      properties:
        object: { type: string, enum: [list] }
        has_more: { type: boolean }
...
    Proveedor:
      type: object
      properties:
        object: { type: string, enum: [proveedor] }
        id: { type: integer }
        nombre: { type: string }
        email: { type: string, format: email }
        telefono: { type: string, nullable: true }
        direccion: { type: string, nullable: true }
        webhook_api_key:
          type: string
          nullable: true
          description: >
            Generada automaticamente al crear el proveedor. Se entrega al proveedor real fuera de
            banda. ADMIN puede consultarla via GET y rotarla con
            PATCH /proveedores/{id}/regenerar-clave-webhook.
        descuento_maximo_pct:
          type: number
          format: double
          minimum: 0
          maximum: 99.99
          description: >
            Descuento maximo (en porcentaje) que se puede negociar sobre el precio de catalogo de
            cualquier producto de este proveedor. 0 junto con aumento_maximo_pct = 0 significa que
            no se negocia (se compra al precio de catalogo).
        aumento_maximo_pct:
          type: number
          format: double
          minimum: 0
          maximum: 100
          description: Aumento maximo (en porcentaje) que se acepta sobre el precio de catalogo.
        activo: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    DetalleRequest:
      type: object
      required: [producto_id, cantidad]
      properties:
        producto_id: { type: integer }
        cantidad: { type: number, format: double, exclusiveMinimum: true, minimum: 0 }
        precio_unitario:
          type: number
          format: double
          minimum: 0.01
          description: >
            Precio unitario negociado (maximo dos decimales). Opcional: omitido, la linea usa el precio de
            catalogo (o, al editar una orden, el que la linea ya tenia). Debe estar entre
            catalogo x (1 - descuento_maximo_pct) y catalogo x (1 + aumento_maximo_pct) del proveedor de la
            orden, ambos extremos incluidos; el catalogo de referencia es el guardado en la linea
            (precio_catalogo), no el vigente hoy. Fuera de rango responde 400 con codigo precio_fuera_de_rango.
    OrdenCompraDetalle:
      type: object
      properties:
        object: { type: string, enum: [orden_compra_detalle] }
        id: { type: integer }
        producto_id: { type: integer }
        producto_codigo: { type: string }
        producto_nombre: { type: string }
        cantidad: { type: number, format: double }
        precio_unitario: { type: number, format: double }
        precio_catalogo:
          type: number
          format: double
          description: Precio de catalogo vigente al crear la linea. Si difiere de precio_unitario, la linea se negocio.
        subtotal: { type: number, format: double }
    OrdenCompraRequest:
      type: object
      required: [proveedor_id, sucursal_destino_id, fecha_necesaria, detalle]
      properties:
        proveedor_id: { type: integer }
        sucursal_destino_id: { type: integer }
        cliente_id:
          type: integer
          nullable: true
          description: Obligatorio si la sucursal destino tiene formato VENTA_DIRECTA
        fecha_necesaria: { type: string, format: date }
        detalle:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/DetalleRequest" }
    OrdenCompra:
      type: object
      properties:
        object: { type: string, enum: [orden_compra] }
        id: { type: integer }
        numero_orden: { type: string, example: OC-2026-000123 }
        estado: { $ref: "#/components/schemas/EstadoOrden" }
        proveedor_id: { type: integer }
        sucursal_destino_id: { type: integer }
        cliente_id: { type: integer, nullable: true }
        cliente_nombre: { type: string, nullable: true }
        fecha_necesaria: { type: string, format: date }
        total: { type: number, format: double }
        motivo_anulacion: { type: string, nullable: true }
        conforme: { type: boolean, nullable: true }
        observacion_cierre: { type: string, nullable: true }
        detalle:
          type: array
          items: { $ref: "#/components/schemas/OrdenCompraDetalle" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        proveedor:
          allOf:
            - $ref: "#/components/schemas/Proveedor"
          nullable: true
          description: Presente solo si la request incluyo ?expand=proveedor; ausente en caso contrario.
...
    LineaTiempoItem:
      type: object
      description: Elemento combinado de auditoría interna o evento de proveedor, discriminado por "origen"
      properties:
        origen: { type: string, enum: [interno, proveedor] }
        tipo: { type: string, description: "estado_nuevo si origen=interno, tipo_evento si origen=proveedor" }
        observacion: { type: string, nullable: true }
        fecha: { type: string, format: date-time }
    ImportacionLoteResponse:
      type: object
      properties:
        object: { type: string, enum: [importacion_lote] }
        ordenes_creadas:
          type: array
          items:
            type: object
            properties:
              referencia_lote: { type: string }
              orden_id: { type: integer }
              numero_orden: { type: string, example: OC-2026-000145 }
        ordenes_rechazadas:
          type: array
          items:
            type: object
            properties:
              referencia_lote: { type: string }
              error:
                type: object
                properties:
                  type: { type: string }
                  code: { type: string }
                  message: { type: string }
                  param: { type: string, nullable: true }